Delete ContentLayout and update AppLayout to allow centered content
