+++ application/account-management/WebApp/routes/admin/account/index.tsx
@@ -7,7 +7,6 @@ import { Trans } from "@lingui/react/macro";
 import { AppLayout } from "@repo/ui/components/AppLayout";
 import { Breadcrumb } from "@repo/ui/components/Breadcrumbs";
 import { Button } from "@repo/ui/components/Button";
-import { ContentLayout } from "@repo/ui/components/ContentLayout";
 import { Form } from "@repo/ui/components/Form";
 import { FormErrorMessage } from "@repo/ui/components/FormErrorMessage";
 import { TextField } from "@repo/ui/components/TextField";
@@ -15,7 +14,7 @@ import { mutationSubmitter } from "@repo/ui/forms/mutationSubmitter";
 import { createFileRoute } from "@tanstack/react-router";
 import { Trash2 } from "lucide-react";
 import { useState } from "react";
-import { Label, Separator } from "react-aria-components";
+import { Separator } from "react-aria-components";
 import DeleteAccountConfirmation from "./-components/DeleteAccountConfirmation";
 
 export const Route = createFileRoute("/admin/account/")({
@@ -35,6 +34,7 @@ export function AccountSettings() {
     <>
       <SharedSideMenu ariaLabel={t`Toggle collapsed menu`} />
       <AppLayout
+        variant="center"
         topMenu={
           <TopMenu>
             <Breadcrumb href="/admin/account">
@@ -46,64 +46,58 @@ export function AccountSettings() {
           </TopMenu>
         }
       >
-        <ContentLayout variant="center">
-          <div className="20 mb-4 flex w-full items-center justify-between space-x-2 sm:mt-4">
-            <div className="mt-3 flex flex-col gap-2 font-semibold text-3xl text-foreground">
-              <h1>
-                <Trans>Account settings</Trans>
-              </h1>
-              <p className="font-normal text-muted-foreground text-sm">
-                <Trans>Manage your account here.</Trans>
-              </p>
-            </div>
-          </div>
+        <h1>
+          <Trans>Account settings</Trans>
+        </h1>
+        <p>
+          <Trans>Manage your account here.</Trans>
+        </p>
 
-          <Form
-            onSubmit={mutationSubmitter(updateCurrentTenantMutation)}
-            validationErrors={updateCurrentTenantMutation.error?.errors}
+        <Form
+          onSubmit={mutationSubmitter(updateCurrentTenantMutation)}
+          validationErrors={updateCurrentTenantMutation.error?.errors}
+          validationBehavior="aria"
+          className="flex flex-col gap-4"
+        >
+          <h2>
+            <Trans>Account information</Trans>
+          </h2>
+          <Separator />
+
+          <Trans>Logo</Trans>
+
+          <img src={logoWrap} alt={t`Logo`} className="max-h-16 max-w-64" />
+          <TextField
+            autoFocus={true}
+            isRequired={true}
+            name="name"
+            defaultValue={tenant?.name ?? ""}
+            isDisabled={updateCurrentTenantMutation.isPending}
+            label={t`Account name`}
             validationBehavior="aria"
-            className="flex flex-col gap-4"
-          >
-            <Label>
-              <Trans>Logo</Trans>
-            </Label>
-            <img src={logoWrap} alt={t`Logo`} className="max-h-16 max-w-64" />
+          />
+          <FormErrorMessage error={updateCurrentTenantMutation.error} />
+          <Button type="submit" className="mt-4" isDisabled={updateCurrentTenantMutation.isPending}>
+            {updateCurrentTenantMutation.isPending ? <Trans>Saving...</Trans> : <Trans>Save changes</Trans>}
+          </Button>
+        </Form>
 
-            <TextField
-              autoFocus={true}
-              isRequired={true}
-              name="name"
-              defaultValue={tenant?.name ?? ""}
-              isDisabled={updateCurrentTenantMutation.isPending}
-              label={t`Account name`}
-              validationBehavior="aria"
-            />
+        <div className="mt-6 flex flex-col gap-4">
+          <h2>
+            <Trans>Danger zone</Trans>
+          </h2>
+          <Separator />
+          <div className="flex flex-col gap-4">
+            <p>
+              <Trans>Delete your account and all data. This action is irreversible—proceed with caution.</Trans>
+            </p>
 
-            <FormErrorMessage error={updateCurrentTenantMutation.error} />
-            <Button type="submit" className="mt-4" isDisabled={updateCurrentTenantMutation.isPending}>
-              {updateCurrentTenantMutation.isPending ? <Trans>Saving...</Trans> : <Trans>Save changes</Trans>}
+            <Button variant="destructive" onPress={() => setIsDeleteModalOpen(true)} className="w-fit">
+              <Trash2 />
+              <Trans>Delete account</Trans>
             </Button>
-          </Form>
-
-          <div className="mt-6 flex flex-col gap-4">
-            <h3 className="font-semibold">
-              <Trans>Danger zone</Trans>
-            </h3>
-            <Separator />
-            <div className="flex flex-col gap-4">
-              <p className="font-normal text-sm">
-                <Trans>Delete your account and all data. This action is irreversible—proceed with caution.</Trans>
-              </p>
-
-              <Button variant="destructive" onPress={() => setIsDeleteModalOpen(true)} className="w-fit">
-                <Trash2 />
-                <Trans>Delete account</Trans>
-              </Button>
-            </div>
           </div>
-
-          <Separator className="my-8" />
-        </ContentLayout>
+        </div>
       </AppLayout>
 
       <DeleteAccountConfirmation isOpen={isDeleteModalOpen} onOpenChange={setIsDeleteModalOpen} />
